refactor: fixed spelling in constants folder
