+++ src/index.ts
@@ -23,7 +23,7 @@ import {
 import erc20Abi from "./abi/erc20.json";
 import { getChainData } from "./utils/getChainData";
 import { getTokenData } from "./utils/getTokenData";
-import { nativeTokenConstant, uint256MaxValue } from "./contants";
+import { nativeTokenConstant, uint256MaxValue } from "./constants";
 
 dotenv.config();
